Benchmark strategy of generating one class per configuration
