+++ subprojects/provider/src/test/kotlin/org/gradle/kotlin/dsl/accessors/AccessorBytecodeEmitterSpike.kt
@@ -16,10 +16,15 @@
 
 package org.gradle.kotlin.dsl.accessors
 
-import com.nhaarman.mockito_kotlin.*
+import com.nhaarman.mockito_kotlin.any
+import com.nhaarman.mockito_kotlin.doReturn
+import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockito_kotlin.only
+import com.nhaarman.mockito_kotlin.verify
 
-import kotlinx.metadata.*
-import kotlinx.metadata.jvm.*
+import kotlinx.metadata.jvm.KotlinClassHeader
+import kotlinx.metadata.jvm.KotlinClassMetadata
+import kotlinx.metadata.jvm.KotlinModuleMetadata
 
 import org.gradle.api.NamedDomainObjectProvider
 import org.gradle.api.Project
@@ -28,8 +33,7 @@ import org.gradle.api.artifacts.ConfigurationContainer
 
 import org.gradle.kotlin.dsl.accessors.AccessorBytecodeEmitter.emitExtensionsMultiThreaded
 import org.gradle.kotlin.dsl.accessors.AccessorBytecodeEmitter.emitExtensionsSingleThreaded
-
-import org.gradle.kotlin.dsl.execution.*
+import org.gradle.kotlin.dsl.accessors.AccessorBytecodeEmitter.emitExtensionsWithOneClassPerConfiguration
 
 import org.gradle.kotlin.dsl.fixtures.TestWithTempFiles
 import org.gradle.kotlin.dsl.fixtures.classLoaderFor
@@ -38,15 +42,14 @@ import org.gradle.kotlin.dsl.fixtures.testCompilationClassPath
 
 import org.gradle.kotlin.dsl.support.compileToDirectory
 import org.gradle.kotlin.dsl.support.loggerFor
+import org.gradle.kotlin.dsl.support.useToRun
 
-import org.jetbrains.org.objectweb.asm.ClassWriter
-import org.jetbrains.org.objectweb.asm.Opcodes
+import org.hamcrest.CoreMatchers.equalTo
+import org.hamcrest.MatcherAssert.assertThat
 
 import org.junit.Test
 
 import java.io.File
-
-import java.util.concurrent.*
 
 import kotlin.concurrent.thread
 
@@ -58,11 +61,36 @@ val ConfigurationContainer.api: NamedDomainObjectProvider<Configuration>
 class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
 
     private
-    val benchmarkConfig = BenchmarkConfig(0, 3)
+    val benchmarkConfig = BenchmarkConfig(0, 5)
 
     @Test
-    fun `benchmark compiled accessors strategy`() {
-        benchmarkWithConfigOnlySchema(::compiledAccessors)
+    fun `verify bytecode accessors (one class per configuration)`() {
+        verifyAccessorsProducedBy { accessors, srcDir, binDir ->
+            val internalNamesOfEmittedClasses =
+                emitExtensionsWithOneClassPerConfiguration(
+                    configOnlySchemaWith(
+                        accessors.filterIsInstance<Accessor.ForConfiguration>().map { it.configurationName }.toList()
+                    ),
+                    srcDir,
+                    binDir
+                )
+            internalNamesOfEmittedClasses.map {
+                it.replace('/', '.')
+            }
+        }
+    }
+
+    @Test
+    fun `verify bytecode accessors (single-threaded)`() {
+        verifyAccessorsProducedBy { accessors, _, binDir ->
+            emitExtensionsSingleThreaded(accessors, binDir)
+            listOf("org.gradle.kotlin.dsl.ConfigurationAccessorsKt")
+        }
+    }
+
+    @Test
+    fun `benchmark bytecode accessors (one class per configuration)`() {
+        benchmarkWithConfigOnlySchema(::bytecodeAccessorsWithOneClassPerConfiguration)
     }
 
     @Test
@@ -73,6 +101,21 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
     @Test
     fun `benchmark bytecode accessors (multi-threaded)`() {
         benchmarkWithConfigOnlySchema(::bytecodeAccessorsMultiThreaded)
+    }
+
+    @Test
+    fun `benchmark compiled accessors strategy (baseline)`() {
+        benchmarkWithConfigOnlySchema(::compiledAccessors)
+    }
+
+    @Test
+    fun `benchmark overhead`() {
+        benchmarkWithConfigOnlySchema { _, _, _ -> }
+    }
+
+    private
+    fun bytecodeAccessorsWithOneClassPerConfiguration(projectSchema: ProjectSchema<String>, srcDir: File, binDir: File) {
+        emitExtensionsWithOneClassPerConfiguration(projectSchema, srcDir, binDir)
     }
 
     private
@@ -94,11 +137,7 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
     }
 
     private
-    fun accessorsForConfigurationsOf(projectSchema: ProjectSchema<String>) =
-        projectSchema.configurations.asSequence().map { Accessor.ForConfiguration(it) }
-
-    private
-    fun benchmarkWithConfigOnlySchema(f: (ProjectSchema<String>, File, File) -> Unit) {
+    fun benchmarkWithConfigOnlySchema(experiment: (ProjectSchema<String>, File, File) -> Unit) {
         var run = 0
         val configOnlySchema = loadConfigurationSchema()
         val benchmarkResult = benchmark(benchmarkConfig) {
@@ -106,7 +145,7 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
                 val outputDir = newFolder("accessors", run.toString(), projectPath.replace(':', '_'))
                 val srcDir = outputDir.resolve("src").apply { mkdir() }
                 val binDir = outputDir.resolve("bin").apply { mkdir() }
-                f(projectSchema, srcDir, binDir)
+                experiment(projectSchema, srcDir, binDir)
                 run += 1
             }
         }
@@ -129,49 +168,57 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
     // it's still not completely fair because of all the overloads
     // in the source code based version but it's enough for a ballpark figure
     private
-    fun loadConfigurationSchema(): Map<String, ProjectSchema<String>> {
-        val helloAndroidProjectSchema = loadMultiProjectSchemaFromResource(
-            "/gradle-project-schema.json"
-//            "/hello-android-project-schema.json"
+    fun loadConfigurationSchema(): Map<String, ProjectSchema<String>> =
+        configOnlySchemaFrom(
+            multiProjectSchemaFromResource(
+//                "/gradle-project-schema.json"
+                "/hello-android-project-schema.json"
+            )
         )
-        return configOnlySchemaFrom(helloAndroidProjectSchema)
+
+    private
+    fun prettyPrint(result: BenchmarkResult) {
+        println("${result.median.ms}ms (${result.points.map { it.ms }})")
     }
 
     private
-    fun prettyPrint(currentResult: BenchmarkResult) {
-        println("${currentResult.median.ms}ms (${currentResult.points.map { it.ms }})")
-    }
+    fun configOnlySchemaFrom(schema: Map<String, ProjectSchema<String>>): Map<String, ProjectSchema<String>> =
+        schema.mapValues { (_, v) -> configOnlySchemaWith(v.configurations) }
 
     private
-    fun configOnlySchemaFrom(schema: Map<String, ProjectSchema<String>>): Map<String, ProjectSchema<String>> = schema.mapValues { (_, v) ->
-        ProjectSchema<String>(
-            configurations = v.configurations,
-            extensions = emptyList(),
-            conventions = emptyList(),
-            tasks = emptyList(),
-            containerElements = emptyList()
-        )
-    }
+    fun configOnlySchemaWith(configurations: List<String>) = ProjectSchema<String>(
+        configurations = configurations,
+        extensions = emptyList(),
+        conventions = emptyList(),
+        tasks = emptyList(),
+        containerElements = emptyList()
+    )
 
     private
-    fun loadMultiProjectSchemaFromResource(resourceName: String): Map<String, ProjectSchema<String>> =
+    fun multiProjectSchemaFromResource(resourceName: String): Map<String, ProjectSchema<String>> =
         loadMultiProjectSchemaFrom(File(javaClass.getResource(resourceName).toURI()))
 
-    @Test
-    fun `spike inlined Configuration accessors`() {
+    private
+    fun verifyAccessorsProducedBy(f: (Sequence<Accessor>, File, File) -> List<String>) {
 
         // given:
-        val accessorsBinDir = newFolder("accessors")
+        val outputDir = newFolder("accessors")
+        val binDir = outputDir.resolve("bin").apply { mkdir() }
+        val srcDir = outputDir.resolve("bin").apply { mkdir() }
+        val accessors = sequenceOf(Accessor.ForConfiguration("api"))
 
         // when:
-        emitExtensionsSingleThreaded(sequenceOf(Accessor.ForConfiguration("api")),
-            outputDir = accessorsBinDir
-        )
+        val classNames = f(accessors, srcDir, binDir)
 
         // then:
-        // verify class
-        classLoaderFor(accessorsBinDir).use {
-            it.loadClass("org.gradle.kotlin.dsl.ConfigurationAccessorsKt").kotlin
+        // verify classes
+        classLoaderFor(binDir).useToRun {
+            classNames.forEach {
+                assertThat(
+                    loadClass(it).kotlin.qualifiedName,
+                    equalTo(it)
+                )
+            }
         }
 
         val configuration = mock<NamedDomainObjectProvider<Configuration>>()
@@ -181,7 +228,7 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
         eval(
             script = "val api = configurations.api",
             target = projectMockWith(configurations),
-            scriptCompilationClassPath = testCompilationClassPath + classPathOf(accessorsBinDir),
+            scriptCompilationClassPath = testCompilationClassPath + classPathOf(binDir),
             baseCacheDir = newFolder("cache")
         )
 
@@ -244,349 +291,5 @@ class AccessorBytecodeEmitterSpike : TestWithTempFiles() {
     private
     fun projectMockWith(configurations: ConfigurationContainer): Project = mock {
         on { getConfigurations() } doReturn configurations
-    }
-}
-
-
-internal
-object AccessorBytecodeEmitter {
-
-    fun emitExtensionsSingleThreaded(accessors: Sequence<Accessor>, outputDir: File) {
-
-        val accessorGetterSignaturePairs = accessors.filterIsInstance<Accessor.ForConfiguration>().map { accessor ->
-            accessor to jvmMethodSignatureFor(accessor)
-        }.toList()
-
-        val metadataWriter = KotlinClassMetadata.FileFacade.Writer()
-        for ((accessor, getterSignature) in accessorGetterSignaturePairs) {
-            metadataWriter.writeMetadataFor(accessor, getterSignature)
-        }
-        val header = metadataWriter.run {
-            (visitExtensions(JvmPackageExtensionVisitor.TYPE) as KmPackageExtensionVisitor).run {
-                visitEnd()
-            }
-            visitEnd()
-            write().header
-        }
-
-        val className = "org/gradle/kotlin/dsl/ConfigurationAccessorsKt"
-        val classBytes =
-            publicClass(className) {
-                visitKotlinMetadataAnnotation(header)
-                for ((accessor, getterSignature) in accessorGetterSignaturePairs) {
-                    method(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, getterSignature.name, getterSignature.desc) {
-                        ALOAD(0)
-                        LDC(accessor.configurationName)
-                        INVOKEINTERFACE(configurationContainerInternalName, "named", namedMethodDescriptor)
-                        ARETURN()
-                    }
-                }
-            }
-
-        outputDir.resolve("$className.class").run {
-            parentFile.mkdirs()
-            writeBytes(classBytes)
-        }
-
-        writeModuleMetadataFor(className, outputDir)
-    }
-
-    fun emitExtensionsMultiThreaded(accessors: Sequence<Accessor>, outputDir: File) {
-
-        val executor = Executors.newFixedThreadPool(2)
-
-        val headerQ = ArrayBlockingQueue<Request>(32)
-        val header = executor.submit(Callable {
-            emitKotlinClassHeader(headerQ)
-        })
-
-        val moduleQ = ArrayBlockingQueue<Request>(32)
-        executor.submit {
-            emitKotlinModule(moduleQ, header, outputDir)
-        }
-
-        accessors.filterIsInstance<Accessor.ForConfiguration>().forEach { accessor ->
-            val r = Request.AccessorOf(accessor, jvmMethodSignatureFor(accessor))
-            headerQ.put(r)
-            moduleQ.put(r)
-        }
-
-        headerQ.put(Request.Done)
-        moduleQ.put(Request.Done)
-
-        executor.shutdown()
-        executor.awaitTermination(1, TimeUnit.DAYS)
-    }
-
-    sealed class Request {
-
-        data class AccessorOf(val accessor: Accessor.ForConfiguration, val signature: JvmMethodSignature) : Request()
-
-        object Done : Request()
-    }
-
-    private
-    fun emitKotlinModule(moduleQ: ArrayBlockingQueue<Request>, header: Future<KotlinClassHeader>, outputDir: File) {
-
-        val className = "org/gradle/kotlin/dsl/ConfigurationAccessorsKt"
-
-        // Write the module metadata
-        writeModuleMetadataFor(className, outputDir)
-
-        val classBytes =
-            publicClass(className) {
-                moduleQ.forEachRequest {
-                    method(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, signature.name, signature.desc) {
-                        ALOAD(0)
-                        LDC(accessor.configurationName)
-                        INVOKEINTERFACE(configurationContainerInternalName, "named", namedMethodDescriptor)
-                        ARETURN()
-                    }
-                }
-                visitKotlinMetadataAnnotation(header.get())
-            }
-
-        outputDir.resolve("$className.class").run {
-            parentFile.mkdirs()
-            writeBytes(classBytes)
-        }
-    }
-
-    private
-    fun writeModuleMetadataFor(className: String, outputDir: File) {
-        val moduleBytes = KotlinModuleMetadata.Writer().run {
-            visitPackageParts("org.gradle.kotlin.dsl", listOf(className), emptyMap())
-            visitEnd()
-            write().bytes
-        }
-        outputDir
-            .resolve("META-INF").apply { mkdir() }
-            .resolve("${outputDir.name}.kotlin_module")
-            .writeBytes(moduleBytes)
-    }
-
-    private
-    fun emitKotlinClassHeader(requests: ArrayBlockingQueue<Request>): KotlinClassHeader {
-
-        val metadataWriter = KotlinClassMetadata.FileFacade.Writer()
-        requests.forEachRequest {
-            metadataWriter.writeMetadataFor(accessor, signature)
-        }
-        return metadataWriter.run {
-            (visitExtensions(JvmPackageExtensionVisitor.TYPE) as KmPackageExtensionVisitor).run {
-                visitEnd()
-            }
-            visitEnd()
-            write().header
-        }
-    }
-
-    private
-    fun KotlinClassMetadata.FileFacade.Writer.writeMetadataFor(accessor: Accessor.ForConfiguration, getterSignature: JvmMethodSignature) {
-        visitProperty(readOnlyPropertyFlags, accessor.configurationName, getterFlags, 6)!!.run {
-            visitReceiverParameterType(0)!!.run {
-                visitClass(configurationContainerInternalName)
-                visitEnd()
-            }
-            visitReturnType(0)!!.run {
-                visitClass(namedDomainObjectProviderInternalName)
-                visitArgument(0, KmVariance.INVARIANT)!!.run {
-                    visitClass(configurationInternalName)
-                    visitEnd()
-                }
-                visitEnd()
-            }
-            (visitExtensions(JvmPropertyExtensionVisitor.TYPE) as JvmPropertyExtensionVisitor).run {
-                visit(null, getterSignature, null)
-                visitSyntheticMethodForAnnotations(null)
-                visitEnd()
-            }
-            visitEnd()
-        }
-    }
-
-    private
-    fun jvmMethodSignatureFor(accessor: Accessor.ForConfiguration): JvmMethodSignature =
-        JvmMethodSignature(
-            "get${accessor.configurationName.capitalize()}",
-            configurationAccessorMethodSignature
-        )
-
-    private
-    val configurationContainerInternalName = ConfigurationContainer::class.internalName
-
-    private
-    val configurationInternalName = Configuration::class.internalName
-
-    private
-    val namedDomainObjectProviderInternalName = NamedDomainObjectProvider::class.internalName
-
-    private
-    val namedMethodDescriptor = "(Ljava/lang/String;)L$namedDomainObjectProviderInternalName;"
-
-    private
-    val configurationAccessorMethodSignature = "(L$configurationContainerInternalName;)L$namedDomainObjectProviderInternalName;"
-
-    private
-    val readOnlyPropertyFlags = flagsOf(
-        Flag.IS_PUBLIC,
-        Flag.Property.HAS_GETTER,
-        Flag.Property.IS_DECLARATION
-    )
-
-    private
-    val getterFlags = flagsOf(
-        Flag.IS_PUBLIC,
-        Flag.PropertyAccessor.IS_NOT_DEFAULT,
-        Flag.PropertyAccessor.IS_INLINE
-    )
-
-    private
-    fun <E> ArrayBlockingQueue<E>.forEachRequest(f: Request.AccessorOf.() -> Unit) {
-        loop@ while (true) {
-            when (val request = take()) {
-                is Request.AccessorOf -> f(request)
-                else -> break@loop
-            }
-        }
-    }
-}
-
-
-internal
-sealed class Accessor {
-
-    data class ForConfiguration(val configurationName: String) : Accessor()
-}
-
-
-/**
- * Writes the given [header] to the class file as a [kotlin.Metadata] annotation.
- **/
-private
-fun ClassWriter.visitKotlinMetadataAnnotation(header: KotlinClassHeader) {
-    visitAnnotation("Lkotlin/Metadata;", true).run {
-        visit("mv", header.metadataVersion)
-        visit("bv", header.bytecodeVersion)
-        visit("k", header.kind)
-        visitArray("d1").run {
-            header.data1.forEach { visit(null, it) }
-            visitEnd()
-        }
-        visitArray("d2").run {
-            header.data2.forEach { visit(null, it) }
-            visitEnd()
-        }
-        visitEnd()
-    }
-}
-
-
-object PrintingVisitor {
-
-    object ForPackage : KmPackageVisitor() {
-
-        override fun visitExtensions(type: KmExtensionType): KmPackageExtensionVisitor? {
-            println("visitExtensions($type)")
-            return object : JvmPackageExtensionVisitor() {
-                override fun visitLocalDelegatedProperty(flags: Flags, name: String, getterFlags: Flags, setterFlags: Flags): KmPropertyVisitor? {
-                    println("visitLocalDelegatedProperty($flags, $name, $getterFlags, $setterFlags)")
-                    return super.visitLocalDelegatedProperty(flags, name, getterFlags, setterFlags)
-                }
-
-                override fun visitEnd() {
-                    println("visitEnd()")
-                    super.visitEnd()
-                }
-            }
-        }
-
-        override fun visitProperty(flags: Flags, name: String, getterFlags: Flags, setterFlags: Flags): KmPropertyVisitor? {
-            println("visitProperty($flags, $name, $getterFlags, $setterFlags)")
-            return ForProperty
-        }
-
-        override fun visitEnd() {
-            println("visitEnd()")
-            super.visitEnd()
-        }
-    }
-
-    object ForProperty : KmPropertyVisitor() {
-
-        override fun visitExtensions(type: KmExtensionType): KmPropertyExtensionVisitor? {
-            println("visitExtensions($type")
-            return object : JvmPropertyExtensionVisitor() {
-                override fun visit(fieldDesc: JvmFieldSignature?, getterDesc: JvmMethodSignature?, setterDesc: JvmMethodSignature?) {
-                    println("visit($fieldDesc, $getterDesc, $setterDesc)")
-                    super.visit(fieldDesc, getterDesc, setterDesc)
-                }
-
-                override fun visitSyntheticMethodForAnnotations(desc: JvmMethodSignature?) {
-                    println("visitSyntheticMethodForAnnotations($desc)")
-                    super.visitSyntheticMethodForAnnotations(desc)
-                }
-
-                override fun visitEnd() {
-                    println("visitEnd()")
-                    super.visitEnd()
-                }
-            }
-        }
-
-        override fun visitReceiverParameterType(flags: Flags): KmTypeVisitor? {
-            println("visitReceiverParameterType($flags)")
-            return ForType
-        }
-
-        override fun visitReturnType(flags: Flags): KmTypeVisitor? {
-            println("visitReturnType($flags)")
-            return ForType
-        }
-
-        override fun visitVersionRequirement(): KmVersionRequirementVisitor? {
-            println("visitVersionRequirement()")
-            return super.visitVersionRequirement()
-        }
-
-        override fun visitEnd() {
-            println("visitEnd()")
-            super.visitEnd()
-        }
-    }
-
-    object ForType : KmTypeVisitor() {
-
-        override fun visitClass(name: ClassName) {
-            println("visitClass($name)")
-            super.visitClass(name)
-        }
-
-        override fun visitArgument(flags: Flags, variance: KmVariance): KmTypeVisitor? {
-            println("visitArgument($flags, $variance)")
-            return ForType
-        }
-
-        override fun visitEnd() {
-            println("visitEnd()")
-            super.visitEnd()
-        }
-    }
-
-    object ForModule : KmModuleVisitor() {
-        override fun visitAnnotation(annotation: KmAnnotation) {
-            println("visitAnnotation($annotation)")
-            super.visitAnnotation(annotation)
-        }
-
-        override fun visitPackageParts(fqName: String, fileFacades: List<String>, multiFileClassParts: Map<String, String>) {
-            println("visitPackageParts($fqName, $fileFacades, $multiFileClassParts")
-            super.visitPackageParts(fqName, fileFacades, multiFileClassParts)
-        }
-
-        override fun visitEnd() {
-            println("visitEnd()")
-            super.visitEnd()
-        }
     }
 }
